+++ ShopperInsights/src/test/java/com/braintreepayments/api/shopperinsights/ShopperInsightsClientUnitTest.kt
@@ -2,6 +2,7 @@ package com.braintreepayments.api.shopperinsights
 
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
+import com.braintreepayments.api.core.AnalyticsEventParams
 import com.braintreepayments.api.core.AnalyticsParamRepository
 import com.braintreepayments.api.core.Authorization
 import com.braintreepayments.api.core.BraintreeClient
@@ -51,23 +52,24 @@ class ShopperInsightsClientUnitTest {
 
     @Test
     fun `when getRecommendedPaymentMethods is called, session id is reset`() {
-        sut.getRecommendedPaymentMethods(mockk(relaxed = true), mockk(relaxed = true))
+        sut.getRecommendedPaymentMethods(mockk(relaxed = true), "some_experiment", mockk(relaxed = true))
 
         verify { analyticsParamRepository.resetSessionId() }
     }
 
     @Test
     fun `when getRecommendedPaymentMethods is called, started event is sent`() {
-        sut.getRecommendedPaymentMethods(mockk(relaxed = true), mockk(relaxed = true))
+        val experiment = "some_experiment"
+        sut.getRecommendedPaymentMethods(mockk(relaxed = true), experiment, mockk(relaxed = true))
 
-        verifyStartedAnalyticsEvent()
+        verifyStartedAnalyticsEvent(AnalyticsEventParams(experiment = experiment))
     }
 
     @Test
     fun `when getRecommendedPaymentMethods is called, failed event is sent`() {
         val request = ShopperInsightsRequest(null, null)
 
-        sut.getRecommendedPaymentMethods(request, mockk(relaxed = true))
+        sut.getRecommendedPaymentMethods(request, "some_experiment", mockk(relaxed = true))
 
         verifyFailedAnalyticsEvent()
     }
@@ -442,15 +444,15 @@ class ShopperInsightsClientUnitTest {
         val apiCallbackSlot = slot<EligiblePaymentsCallback>()
         every { api.findEligiblePayments(any(), capture(apiCallbackSlot)) } just runs
 
-        sut.getRecommendedPaymentMethods(request, callback)
+        sut.getRecommendedPaymentMethods(request, "some_experiment", callback)
 
         apiCallbackSlot.captured.onResult(result = result, error = error)
     }
 
-    private fun verifyStartedAnalyticsEvent() {
+    private fun verifyStartedAnalyticsEvent(params: AnalyticsEventParams = AnalyticsEventParams()) {
         verify {
             braintreeClient
-                .sendAnalyticsEvent("shopper-insights:get-recommended-payments:started")
+                .sendAnalyticsEvent("shopper-insights:get-recommended-payments:started", params)
         }
     }